gconf-client: Use the cache for keys we know we don't have locally
Say the application recursively preloads /foo which contains /foo/a/a and
/foo/b/b, which are fetched and cached.  The application then asks for /foo/c/c.
Because we recursively preloaded we should know that this key doesn't exist, but
because it isn't stored in the caches we have to make a DBus call.

This change lets the caching layer know that we recursively loaded /foo and
therefore if the key isn't in the cache, it doesn't exist.