introspection: Add annotations to fix warnings
Add (skip) (transfer) and (element-type) annotations to fix
almost all warnings from introspection (there is one bug left in
the scanner causing a warning about GConfError.)

Docs for the functions being annotated are moved to the source files.

All types that are not currently boxed or GObject are skipped; some
of these could be made boxed with only a small amount of work.