gconf-dbus: On SIGHUP, do not recreate databases but reload the sources
When we destroy and recreate a database, the object path associated to
it on dbus changes. This is bad as clients do not know about the object
path change, and therefore fail.

Instead of destroying and recreating the databases, we simply reload the
sources for this database. This achieves the same result, without
changing the object path.

Note that for the corba backend, we still use the old way as simply
reloading the sources might not work -- the sources are used in
corba-specific code.

https://bugzilla.gnome.org/show_bug.cgi?id=659835